# fix(native filters): rendering performance improvement by reduce overrendering