--- a/superset-frontend/src/dataMask/reducer.ts
+++ b/superset-frontend/src/dataMask/reducer.ts
@@ -56,7 +56,6 @@ export function getInitialDataMask(
   }
   return {
     ...otherProps,
-    __cache: {},
     extraFormData: {},
     filterState: {},
     ownState: {},