### SUMMARY
Following up #25282, rendering performance can still be improved by reducing unnecessary rerenders (also known as over-rendering).
This commit improves performance in four areas.

#### 1. Inefficient uniqWith and object equality check in uniqueOptions

https://github.com/apache/superset/blob/d6fde3cdd5987e3bf9a37d1fcde2dc82ef695747/superset-frontend/src/filters/components/Select/SelectFilterPlugin.tsx#L221-L224

![Screenshot 2023-11-01 at 1 48 34 PM](https://github.com/apache/superset/assets/1392866/5d619acc-9001-4338-87b8-caa94eda018d)

To get the list of unique values, it compares the entire data list (`arrayIncludesWith`) using object equality comparison (`isEqual`). This is slow because the data contains several attributes, and object equality comparison is slow.

As it shown in the above metric, `uniqueOptions` operation triggers four times every time a filter component re-renders, which takes more than 2 seconds per re-render for 10 filters (i.e. 50ms * 4 * 10 filters = 2s).

Using a hashmap approach (with unnecessary equality check)  reduced the delay from ~200m(=50ms * 4) to 10ms (95% reduction)
![Screenshot 2023-11-01 at 1 47 15 PM](https://github.com/apache/superset/assets/1392866/bd271944-c87d-4937-ade8-106fbc568489)

#### 2. Redundant filter re-renders due to the dynamic sizing

|Initially renders at a minimum width and then expands to its full width.|
|--|
|![native-filter-width-overrendering](https://github.com/apache/superset/assets/1392866/ea9f768b-dd3a-4f25-9e9e-fd68005f713f)|

Since the native filter's allocated to 100% of its width using [SuperChart](https://github.com/apache/superset/blob/e7a187680713867f22b082f3bb0a57296d2a331c/superset-frontend/packages/superset-ui-core/src/chart/components/SuperChart.tsx#L245-L258), which is wrapped by `@vx/responsive`, it triggers re-renders every time the screen size is updated.
This commit removes the resize observer (by assigning the static width) and updates the filter component wrapper to the full width using CSS instead. As a result, the filters no longer re-render when the page size changes.
![native-filter-css-width](https://github.com/apache/superset/assets/1392866/19fc445b-4a2f-4808-a60d-b7c94a2d382b)

#### 3. descendants over-rendering

React's inherent behavior of rerendering all descendant components, coupled with the Redux container's rerendering upon state changes, leads to excessive rerendering of dashboard components, including filter components, whenever the chart fetch status updates the Redux store.
This commit replaces DashboardBuilder to the identical descendant (`props.children`) which prevents the rerenders when the container component rerenders.
To further optimize performance, the dashboard state synchronization logic was moved to a separate component, effectively decoupling it from the DashboardPage component and preventing unnecessary rerenders of the DashboardPage.

#### 4. Redundant `__cache` attribute

|Due to superfluous setDataMask calls triggered by __cache update, each setState update causes four unnecessary uniqWith executions|
|--|
|![fix_native_filters___rendering_performance_improvement_by_reduce_overrendering_by_justinpark_·_Pull_Request__25901_·_apache_superset](https://github.com/apache/superset/assets/1392866/f9df28ce-f6fb-4bd9-9aed-a817a0bbf064)|

An unnecessary __cache value was introduced in the SelectFilterPlugin in [PR#14869](https://github.com/apache/superset/pull/14869/commits/1749ed542a4f4cedc6847316bf65fc2b82eb1fbd#diff-b4d2de68c8a3b04879772237085ede1b49a6908b6f2890a9ab1aaeae9e18b854R65), but it was never utilized and resulted in superfluous rerenders of filter components.
This commit removes this redundant `__cache` value to resolve the problem.

### BEFORE/AFTER SCREENSHOTS OR ANIMATED GIF

Successfully optimized the application's performance, resulting in a 10% decrease in rendering count (255 to 228) and a remarkable 60% reduction in loading time (10.9 seconds to 4.5 seconds).

|Before|After|
|--|--|
|![Screenshot 2023-11-07 at 3 57 28 PM](https://github.com/apache/superset/assets/1392866/2e5dc6ad-a5e1-4e66-9e96-01f1fb3d7733)|![Screenshot 2023-11-07 at 3 58 18 PM](https://github.com/apache/superset/assets/1392866/4d82aa63-b884-467c-ace7-b2834c2d2841)|


### TESTING INSTRUCTIONS

1. Go to any dashboard
2. Add more than 10 select filters
3. Reload dashboard
4. See that filters take a lot of time for before render.

### ADDITIONAL INFORMATION
<!--- Check any relevant boxes with "x" -->
<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->
- [ ] Has associated issue:
- [ ] Required feature flags:
- [ ] Changes UI
- [ ] Includes DB Migration (follow approval process in [SIP-59](https://github.com/apache/superset/issues/13351))
  - [ ] Migration is atomic, supports rollback & is backwards-compatible
  - [ ] Confirm DB migration upgrade and downgrade tested
  - [ ] Runtime estimates and downtime expectations provided
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API