@justinpark 2023-11-17
> Thanks jinghua-qa! justinpark Would you rather address my comments in this PR or merge it and do a follow up if needed?

@kgabryje please see my comments.